Rewrite finder as a flask app.
Really, this is just a restructuring. I've used the structure of a
different project I made called dash, which was probably based off
of the structure I made for sidechannel, which was itself based off
of an earlier structure and so on. It leads back to the combined
advice of Miguel Grinberg[1] and the dozen Flask apps or so I've
seen at work.

As you can see from the change-list, I mostly pulled the assets
straight into a static directory, and made index.html into a
template in that it refers to them using jinja instead of a direct
url.

At this import, I'm stopping at "it shows pictures and the JS and
CSS seem to work." The image-posting is not going to work yet --
I will write the image model next and get it back, then add a list
of the "best" images so far. And then we'll move on to users and
personal lists.

1 - http://blog.miguelgrinberg.com/post/the-flask-mega-tutorial-part-i-hello-world